backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
sqlalchemy==2.0.27
pydantic==2.6.1
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1 